Some parallel options
- wabbit objects can now read in blocks one by one instead of all data at the same time
- hdf2vtkhdf has gone parallel, parallel execution with mpirun, but the performance is not really extraordinary
- hdf2vtkhdf now reads in the blocks one by one so I expect it to work with any filesize

+++ bin/hdf2vtkhdf.py
@@ -1,5 +1,14 @@
 #!/usr/bin/env python3
 import h5py, os, sys, argparse, glob, time, numpy as np
+try:
+  from mpi4py import MPI
+  mpi_size = MPI.COMM_WORLD.Get_size()
+  mpi_parallel = mpi_size > 1
+  mpi_rank = MPI.COMM_WORLD.Get_rank()
+except:
+  mpi_parallel = False
+  mpi_rank = 0
+  mpi_size = 1
 
 sys.path.append(os.path.join(os.path.split(__file__)[0], ".."))
 import wabbit_tools
@@ -85,11 +94,11 @@ def hdf2vtkhdf(w_obj: wabbit_tools.WabbitHDF5file, save_file=None, verbose=True,
   # check if vectors are full elsewise add them as scalars
   for pre in v_names:
     if (pre+'x' in p_names and pre+'y' in p_names and pre+'z' in p_names and w_main.dim==3):
-        if verbose: print( pre+' is a 3D vector (x,y,z)')
+        if verbose and mpi_rank==0: print( f'   {pre} is a 3D vector (x,y,z)')
     elif (pre+'x' in p_names and pre+'y' in p_names and w_main.dim==2):
-        if verbose: print( pre+' is a 2D vector (x,y)')
+        if verbose and mpi_rank==0: print( f'   {pre} is a 2D vector (x,y)')
     else:
-        print( f"WARRNING: {pre} is not a vector (its x-, y- or z- component is missing..)")
+        if mpi_rank==0: print( f"   WARRNING: {pre} is not a vector (its x-, y- or z- component is missing..)")
         v_ind.remove(v_ind[v_names.index(pre)])
         v_names.remove( pre )
         # if pre+'x' in p_names: scalars.append(pre+'x')
@@ -101,38 +110,40 @@ def hdf2vtkhdf(w_obj: wabbit_tools.WabbitHDF5file, save_file=None, verbose=True,
   if save_file is None: save_file = w_main.orig_file.replace(".h5", file_ending)
   if not save_file.endswith(file_ending): save_file += file_ending
 
-  # Create the HDF5 file with main group
-  if os.path.isfile(save_file): os.remove(save_file)  # a bit brute-force, maybe ask for deletion?
-  f =  h5py.File(save_file, 'w')
+  # host deletes old file if it exists
+  if mpi_rank == 0:
+    if os.path.isfile(save_file): os.remove(save_file)  # a bit brute-force, maybe ask for deletion?
+  if mpi_parallel: MPI.COMM_WORLD.Barrier()  # ensure that file is properly deleted before all processes continue#
+
+  # now all processes open the file
+  if not mpi_parallel: f =  h5py.File(save_file, 'w')
+  else: f = h5py.File(save_file, 'w', driver='mpio', comm=MPI.COMM_WORLD)
+
   vtkhdf_group = f.create_group('VTKHDF', track_order=True)
   vtkhdf_group.attrs.create('Type', np.string_('PartitionedDataSetCollection'))
   vtkhdf_group.attrs.create('Version', np.array([2, 3], dtype='i8'))
   assembly_group = vtkhdf_group.create_group('Assembly')
 
-  # now add every block
+  # collective loop creating the metadata - all processes need to do this
   start_time = time.time()
   for i_block in range(w_main.total_number_blocks):
-    rem_time = (w_main.total_number_blocks - i_block) * (time.time() - start_time) / (i_block + 1e-4*(i_block == 0))
-    # Format remaining time in HH:MM:SS format
-    hours, rem = divmod(rem_time, 3600)
-    minutes, seconds = divmod(rem, 60)
-    if verbose:
-        print_progress_bar(i_block, w_main.total_number_blocks, prefix=f'Processing:', suffix=f'ETA: {int(hours):02d}h {int(minutes):02d}m { seconds:02.1f}s')
+    # for overfull CVS grids we have the option to split them into levels to make the overlay visible
+    split_levels_add = (split_levels * (w_main.level[i_block]-1) * np.max(w_main.domain_size))
 
     # Create this block itself
     block_group = vtkhdf_group.create_group(f'Block{i_block}')
 
     # Add attributes to block
     if w_main.dim == 2:
-        block_group.attrs.create('Direction', np.array([1, 0, 0, 0, 1, 0, 0, 0, 1], dtype='f8'))
-        block_group.attrs.create('Origin', np.append(w_main.coords_origin[i_block][::-1], 0), dtype='f8')
-        block_group.attrs.create('Spacing', np.append(w_main.coords_spacing[i_block][::-1], 0), dtype='f8')
-        block_group.attrs.create('WholeExtent', np.array([0, w_main.block_size[0]-1, 0, w_main.block_size[1]-1, 0, 1], dtype='i8'))      
+      block_group.attrs.create('Direction', np.array([1, 0, 0, 0, 1, 0, 0, 0, 1], dtype='f8'))
+      block_group.attrs.create('Origin', np.append(w_main.coords_origin[i_block][::-1], 0) + np.array([0,0,split_levels_add]), dtype='f8')
+      block_group.attrs.create('Spacing', np.append(w_main.coords_spacing[i_block][::-1], 0), dtype='f8')
+      block_group.attrs.create('WholeExtent', np.array([0, w_main.block_size[0]-1, 0, w_main.block_size[1]-1, 0, 1], dtype='i8'))      
     else:
-        block_group.attrs.create('Direction', np.array([1, 0, 0, 0, 1, 0, 0, 0, 1], dtype='f8'))
-        block_group.attrs.create('Origin', w_main.coords_origin[i_block][::-1], dtype='f8')
-        block_group.attrs.create('Spacing', w_main.coords_spacing[i_block][::-1], dtype='f8')
-        block_group.attrs.create('WholeExtent', np.array([0, w_main.block_size[0]-1, 0, w_main.block_size[1]-1, 0, w_main.block_size[2]-1], dtype='i8'))
+      block_group.attrs.create('Direction', np.array([1, 0, 0, 0, 1, 0, 0, 0, 1], dtype='f8'))
+      block_group.attrs.create('Origin', w_main.coords_origin[i_block][::-1] + np.array([0,0,split_levels_add]), dtype='f8')
+      block_group.attrs.create('Spacing', w_main.coords_spacing[i_block][::-1], dtype='f8')
+      block_group.attrs.create('WholeExtent', np.array([0, w_main.block_size[0]-1, 0, w_main.block_size[1]-1, 0, w_main.block_size[2]-1], dtype='i8'))
     block_group.attrs.create('Type', np.string_('ImageData'))
     block_group.attrs.create('Version', np.array([2, 3], dtype='i8'))
     block_group.attrs.create('Index', i_block, dtype='i8')
@@ -141,22 +152,52 @@ def hdf2vtkhdf(w_obj: wabbit_tools.WabbitHDF5file, save_file=None, verbose=True,
     # Add block data
     cell_data_group = block_group.create_group('CellData')
 
+    # Create dataset for scalars
+    for i_s, i_n in zip(s_names, s_ind):
+      if w_main.dim == 2: cell_data_group.create_dataset(i_s, shape=[1,w_main.block_size[0]-1, w_main.block_size[1]-1], dtype=np.float64)
+      else: cell_data_group.create_dataset(i_s, shape=w_main.block_size-1, dtype=np.float64)
+    # Create datasets for vectors
+    for i_v, i_n in zip(v_names, v_ind):
+      if w_main.dim == 2: cell_data_group.create_dataset(i_v, shape=[1,w_main.block_size[0]-1, w_main.block_size[1]-1, w_main.dim], dtype=np.float64)
+      else: cell_data_group.create_dataset(i_v, shape=np.append(w_main.block_size-1, w_main.dim), dtype=np.float64)
+  if args.verbose and mpi_rank == 0: print(f"   Created metadata: {time.time() - start_time:.3f} seconds")
+
+  # independent loop attaching the data - this is parallelized
+  start_time = time.time()
+  for i_block in range(int(mpi_rank/mpi_size*w_main.total_number_blocks), int((mpi_rank+1)/mpi_size*w_main.total_number_blocks)):
+    rem_time = (w_main.total_number_blocks - i_block) * (time.time() - start_time) / (i_block + 1e-4*(i_block == 0))
+    # Format remaining time in HH:MM:SS format
+    hours, rem = divmod(rem_time, 3600)
+    minutes, seconds = divmod(rem, 60)
+    if verbose and mpi_rank==0 and i_block < int(w_main.total_number_blocks/mpi_size):
+        print_progress_bar(i_block, int(w_main.total_number_blocks/mpi_size), prefix=f'   Processing data:', suffix=f'ETA: {int(hours):02d}h {int(minutes):02d}m { seconds:02.1f}s')
+
+    # get celldatagroup
+    cell_data_group = vtkhdf_group[f'Block{i_block}']['CellData']
+
     # Attach data for scalars - currently copying but maybe there is a more clever way
     for i_s, i_n in zip(s_names, s_ind):
-        if w_main.dim == 2: cell_data_group.create_dataset(i_s, data=w_obj_list[i_n].blocks[i_block, :-1, :-1].reshape([1,w_main.block_size[0]-1, w_main.block_size[1]-1]), dtype=np.float64)
-        else: cell_data_group.create_dataset(i_s, data=w_obj_list[i_n].blocks[i_block, :-1, :-1, :-1], dtype=np.float64)
+      if w_main.dim == 2:
+        data_append = w_obj_list[i_n].block_read(i_block)[:-1, :-1].reshape([1,w_main.block_size[0]-1, w_main.block_size[1]-1])
+      else:
+        data_append = w_obj_list[i_n].block_read(i_block)[:-1, :-1, :-1]
+      cell_data_group[i_s][:] = data_append
     # Attach data for vectors - currently copying but maybe there is a more clever way
     for i_v, i_n in zip(v_names, v_ind):
-        if w_main.dim == 2:
-          data_append = np.zeros([1, w_main.block_size[0]-1, w_main.block_size[1]-1, w_main.dim])
-          for i_ind, i_ndim in enumerate(i_n): data_append[0,:,:,i_ind] = w_obj_list[i_ndim].blocks[i_block, :-1, :-1]
-        else:
-          data_append = np.zeros([w_main.block_size[0]-1, w_main.block_size[1]-1, w_main.block_size[2]-1, w_main.dim])
-          for i_ind, i_ndim in enumerate(i_n): data_append[:,:,:,i_ind] = w_obj_list[i_ndim].blocks[i_block, :-1, :-1, :-1]
-        cell_data_group.create_dataset(i_v, data=data_append, dtype=np.float64)
+      if w_main.dim == 2:
+        data_append = np.zeros([1, w_main.block_size[0]-1, w_main.block_size[1]-1, 0])
+        for i_ndim in i_n: data_append = np.append(data_append, w_obj_list[i_ndim].block_read(i_block)[np.newaxis, :-1, :-1, np.newaxis], axis=3)
+      else:
+        data_append = np.zeros([w_main.block_size[0]-1, w_main.block_size[1]-1, w_main.block_size[2]-1, 0])
+        for i_ndim in i_n: data_append = np.append(data_append, w_obj_list[i_ndim].block_read(i_block)[:-1, :-1, :-1, np.newaxis], axis=3)
+      cell_data_group[i_v][:] = data_append
 
   # close file
   f.close()
+
+  if args.verbose and mpi_rank == 0: print(f"   Added data:       {time.time() - start_time:.3f} seconds")
+
+  if mpi_parallel: MPI.Finalize()
             
 
 
@@ -194,10 +235,14 @@ if __name__ == "__main__":
   # group3.add_argument("-l", "--skip-incomplete-prefixes", help="If some files are missing, skip the prefix", action="store_true")
   args = parser.parse_args()
 
-  if args.verbose:
-    print( bcolors.OKGREEN + "**********************************************" + bcolors.ENDC )
-    print( bcolors.OKGREEN + "**   hdf2xml.py                             **" + bcolors.ENDC )
-    print( bcolors.OKGREEN + "**********************************************" + bcolors.ENDC )
+  if args.verbose and mpi_rank == 0:
+    print( bcolors.OKGREEN + "*"*50 + bcolors.ENDC )
+    if not mpi_parallel:
+      print( bcolors.OKGREEN + "**    " + f'hdf2vtkhdf.py in serial mode'.ljust(42) + "**" + bcolors.ENDC )
+    else:
+      print( bcolors.OKGREEN + "**    " + f'hdf2vtkhdf.py in parallel mode, np={mpi_size}'.ljust(42) + "**" + bcolors.ENDC )
+
+    print( bcolors.OKGREEN + "*"*50 + bcolors.ENDC )
   
   # set directory in case infile is dir and outfile is default
   if args.outfile == "all" and os.path.isdir(args.infile):
@@ -207,7 +252,7 @@ if __name__ == "__main__":
   time_process = {}
   if os.path.isfile(args.infile) and args.infile.endswith(".h5"):
     state_1 = wabbit_tools.WabbitHDF5file()
-    state_1.read(args.infile, verbose=args.verbose)
+    state_1.read(args.infile, read_var="meta", verbose=args.verbose and mpi_rank == 0)
     time_1 = np.round(state_1.time, 12)  # round to 12 digits to avoid floating points diffrences
     time_process[time_1] = [state_1]
     filelist = [1]  # for verbose
@@ -218,7 +263,7 @@ if __name__ == "__main__":
     filelist = sorted( glob.glob(os.path.join(args.infile,"*.h5")) )
     for i_file in filelist:
       state_1 = wabbit_tools.WabbitHDF5file()
-      state_1.read(i_file, verbose=args.verbose)
+      state_1.read(i_file, verbose=args.verbose and mpi_rank == 0)
       time_1 = np.round(state_1.time, 12)  # round to 12 digits to avoid floating points diffrences
       if not time_1 in time_process:
         time_process[time_1] = []
@@ -226,11 +271,15 @@ if __name__ == "__main__":
   
   if len(time_process) == 0:
     print(bcolors.FAIL + f"ERROR: I did not find any .h5 files on path {args.infile}" + bcolors.ENDC)
-  if args.verbose:
+  if args.verbose and mpi_rank == 0:
     print(f"Found {len(filelist)} .h5 file(s) on {len(time_process)} time instant(s)")
 
   for i_n, i_time in enumerate(time_process):
-    if args.verbose: print(f"Time {i_time}, {i_n+1}/{len(time_process)}")
+    start_time = time.time()
+    if args.verbose and mpi_rank == 0: print(f"Time {i_time}, {i_n+1}/{len(time_process)}")
 
     # create vtkhdf
     hdf2vtkhdf(time_process[i_time], save_file=f"{args.outfile}_{wabbit_tools.time2wabbitstr(i_time)}", verbose=args.verbose, scalars=args.scalars, split_levels=args.cvs_split_levels)
+
+    # output timing
+    if args.verbose and mpi_rank == 0: print(f"   Converted file:   {time.time() - start_time:.3f} seconds")
